+++ tests/backend/Feature/AuthTest.php
@@ -39,6 +39,39 @@ class AuthTest extends TestCase
         $this->assertUnprocessable();
     }
 
+    public function testBruteForceLogin()
+    {
+        $this->sendRequest('POST', '/login', [
+            'username' => 'fake',
+            'password' => 'fake',
+        ], [], ['REMOTE_ADDR' => '10.10.10.10']);
+        $this->assertUnprocessable();
+
+        for ($i = 0; $i < 20; $i++) {
+            $this->sendRequest('POST', '/login', [
+                'username' => 'fake',
+                'password' => 'fake',
+            ], [], ['REMOTE_ADDR' => '10.10.10.10']);
+        }
+        $this->assertStatus(429);
+
+        for ($i = 0; $i < 20; $i++) {
+            $this->sendRequest('POST', '/login', [
+                'username' => 'fake',
+                'password' => 'fake',
+            ], [], ['REMOTE_ADDR' => '2001:db8:3333:4444:5555:6666:7777:8888']);
+        }
+        $this->assertStatus(429);
+
+
+        // new ip should be ok
+        $this->sendRequest('POST', '/login', [
+            'username' => 'fake',
+            'password' => 'fake',
+        ], [], ['REMOTE_ADDR' => '10.10.10.1']);
+        $this->assertUnprocessable();
+    }
+
     public function testAlreadyLoggedIn()
     {
         $username = 'john@example.com';